posting change to Compoent and implemented edit mode posting window
