+++ srcs/front/css/main.css
@@ -698,6 +698,15 @@ a {
 	height: 10vmin;
 }
 
+.posting-background .App-logo {
+	position: absolute;
+	top: 50%;
+	left: 50%;
+	z-index: inherit;
+	transform: translate(-50%, -50%);
+	animation: App-logo-spin infinite 14s linear;
+}
+
 .App-header {
 	/* background-color: #282c34; */
 	/* background-image: url(../public/images/bg.jpg); */
@@ -994,6 +1003,7 @@ ul {
 
 #post-container {
 	position: absolute;
+	min-height: 100vh;
 	width: 100%;
 	top: 0;
 	z-index: 1;
@@ -1004,6 +1014,7 @@ ul {
 #post-main {
 	display: block;
 	margin-top: 160px;
+
 }
 
 .post-inner-index {
@@ -1148,8 +1159,33 @@ ul {
 	line-height: 28px;
 }
 
+.posting-background {
+	background: rgba(0, 0, 0, 0.2);
+	position: absolute;
+	top: 0;
+	left: 0;
+	width: 100vw;
+	min-height: 100vh;
+	z-index: 300;
+}
+
+.posting-wrapper {
+	position: relative;
+	width: 90%;
+	height: auto;
+	margin: 2% 5%;
+	background: white;
+	padding: 0;
+	border-radius: 20px;
+	box-shadow: 0px 0px 8px 4px rgba(90, 90, 90, 0.3);
+}
+
 .posting-wrap {
-	margin-bottom: 50px;
+	margin: 5%;
+}
+
+.posting-container {
+	padding: 20px 0;
 }
 
 .posting-btn {
@@ -1168,7 +1204,7 @@ ul {
 	justify-content: center;
 	font-size: 20px;
 	transition: 0.3s;
-	z-index: 999;
+	z-index: 100;
 }
 
 .posting-btn:hover {
@@ -1318,7 +1354,23 @@ ul {
 	transition: 0.3s;
 }
 
+.submit-wrap .posting-close-btn {
+	position: absolute;
+	right: 100px;
+	top: 50%;
+	width: 100px;
+	height: 38px;
+	background: #333;
+	color: white;
+	cursor: pointer;
+	border-radius: 20px;
+	transition: 0.3s;
+}
+
 .submit-wrap .post-submit-btn:hover {
+	background: #f55;
+}
+.submit-wrap .posting-close-btn:hover {
 	background: #f55;
 }
